feat(telemetry): add WASM support for Otel
